% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fetcher.R
\name{connect.database}
\alias{connect.database}
\title{Opens a connection to the SQL server, and optionally connects to the
database.}
\usage{
connect.database(dbname = NULL)
}
\arguments{
\item{dbname}{The name of the database. A complete list of names can be found
using show.databases after running connect.database().}
}
\description{
Opens a connection to the SQL server, and optionally connects to the
database.
}
\examples{
connect.database("drugs_agora")
}

