% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/process.R
\name{construct.index}
\alias{construct.index}
\title{Constructs a price index by binning price data.}
\usage{
construct.index(market = "agora", category = "2361707", units = NULL,
  additionalQuery = "", binsize = 1000, scale = 1)
}
\arguments{
\item{market}{The name of the marketplace to draw data from.}

\item{category}{The category of drug to use.}

\item{units}{Units the drug is measured in.}

\item{additionalQuery}{Any additional SQL constraints on the price data.}

\item{scale}{Optional scaling factor.}

\item{dates}{The dates which form the endpoints of the bins. If a single
number, then the full date range is used, and evenly split according to the
number.}
}
\description{
Constructs a price index by binning price data.
}
\examples{
construct.index(market = "agora",
                category = "2361707",
                units = "mg",
                additionalQuery = "amount > 50",
                dates = 20)
}

