% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fetcher.R
\name{download.stride}
\alias{download.stride}
\title{Downloads from the STRIDE database. Supports cocaine and meth, the only two datasets publicly available.}
\usage{
download.stride(category = "cocaine")
}
\arguments{
\item{category}{Either 'cocaine' or 'meth'.}
}
\description{
Downloads from the STRIDE database. Supports cocaine and meth, the only two datasets publicly available.
}

