% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fetcher.R
\name{get.connection}
\alias{get.connection}
\title{Get database connection from the package environment}
\usage{
get.connection()
}
\description{
Get database connection from the package environment
}

