% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mysql.R
\name{get.date.range}
\alias{get.date.range}
\title{Returns the date that scraping began and ended for this marketplace.}
\usage{
get.date.range()
}
\description{
Returns the date that scraping began and ended for this marketplace.
}

