% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fetcher.R
\name{get.selected.database}
\alias{get.selected.database}
\title{Returns the database currently connected to.}
\usage{
get.selected.database()
}
\description{
Returns the database currently connected to.
}

