% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mysql.R
\name{select.database}
\alias{select.database}
\title{Connects to the database.}
\usage{
select.database(dbname)
}
\arguments{
\item{dbname}{The name of the database. A complete list of names can be found
using show.databases after running connect.database().}
}
\description{
Connects to the database.
}
\examples{
select.database("drugs_agora")
}

